add dark mode toggle and cross-platform support
- theme: add css variables, data-theme dark mode, and toggle button

- platform: add include/platform.h for windows/mingw compatibility

- platform: fix %zu format specifiers, mkdir, access, socket apis

- platform: replace fork/exec with cxo_spawnvp, add winsock support

- tests: all pass on windows mingw

+++ src/linker.c
@@ -104,7 +104,7 @@ int cxo_link_entries(cxo_context_t* ctx, arena_t* arena)
         cxo_entry_t* existing;
         
         if (!entry->id) {
-            fprintf(stderr, "Warning: entry %zu has no id, skipping\n", i);
+            fprintf(stderr, "Warning: entry %lu has no id, skipping\n", (unsigned long)i);
             continue;
         }
