+++ tests/test_scanner.c
@@ -38,15 +38,15 @@ int main(void)
     printf("PASS: cxo_scan_content\n");
     
     /* Verify results */
-    printf("\nFound %zu entries:\n", ctx->count);
+    printf("\nFound %lu entries:\n", (unsigned long)ctx->count);
     for (size_t i = 0; i < ctx->count; i++) {
         cxo_entry_t* e = ctx->entries[i];
-        printf("  [%zu] id=%s, lang=%s, slug=%s, path=%s\n",
+        printf("  [%lu] id=%s, lang=%s, slug=%s, path=%s\n",
                i, e->id, e->lang, e->slug, e->md_content);
     }
     
     if (ctx->count < 2) {
-        printf("\nFAIL: Expected at least 2 entries, got %zu\n", ctx->count);
+        printf("\nFAIL: Expected at least 2 entries, got %lu\n", (unsigned long)ctx->count);
         arena_destroy(arena);
         return 1;
     }
@@ -57,7 +57,7 @@ int main(void)
         char expected_path[256];
         
         if (!e->id || !e->lang || !e->slug || !e->md_content) {
-            printf("FAIL: Entry %zu has NULL fields\n", i);
+            printf("FAIL: Entry %lu has NULL fields\n", (unsigned long)i);
             arena_destroy(arena);
             return 1;
         }
@@ -65,14 +65,14 @@ int main(void)
         snprintf(expected_path, sizeof(expected_path), "content/%s/%s.md",
                  e->lang, e->slug);
         if (strcmp(e->md_content, expected_path) != 0) {
-            printf("FAIL: Entry %zu path mismatch: expected %s, got %s\n",
+            printf("FAIL: Entry %lu path mismatch: expected %s, got %s\n",
                    i, expected_path, e->md_content);
             arena_destroy(arena);
             return 1;
         }
         
         if (strcmp(e->id, e->slug) != 0) {
-            printf("FAIL: Entry %zu id != slug: %s vs %s\n",
+            printf("FAIL: Entry %lu id != slug: %s vs %s\n",
                    i, e->id, e->slug);
             arena_destroy(arena);
             return 1;
@@ -80,7 +80,7 @@ int main(void)
     }
     printf("PASS: All entry paths and fields correct\n");
     
-    printf("\nPASS: Found expected %zu entries\n", ctx->count);
+    printf("\nPASS: Found expected %lu entries\n", (unsigned long)ctx->count);
     
     arena_destroy(arena);
     printf("PASS: cleanup\n");
